std.time.epoch: support any timestamp (negative seconds)
There's 3 things keeping std.time.epoch from supporting any timestamp,
Year which is hardcoded to u16, EpochSeconds which uses u64 and EpochDay
which uses u47 (bit width calculated to accomodate any EpochSeconds).

To address the `Year` limitation we can remove the `Year` alias that was
defined to `u16`. Functions can simply take `anytype` for the year. For
`YearAndDay` which has a year field, we'll make it a function that returns
a struct given the year type.

For EpochSeconds/EpochDay structs, we'll turn them into functions that
take the underlying integer type and return structs instead. Typical usage
is to create an EpochSeconds type and use that to create an EpochDay, so
we'll calculate the type for EpochDay based on the type for seconds.

The previous implementation only worked with unsigned/positive integers
so I had to add a bit of logic to handle negative seconds/days.  I've also
added tests for the "negative cases", including the "Windows Epoch" which
goes all the way back to 1601.

I think having these utilities take the underlying integer type as a
generic parameter makes sense because the typical usage for these types
will be to instantiate them based on a timestamp provided by the
underlying platform.  Each platform will already have a type defined
to represent the timestamp so all that's needed is to use that same time
in these generic functions.